Evicting load and all the from_* functions logic The load and all the from_* functions now reside at mbuild/conversion.py Still need to resolve some bugs. Will also considering moving the save() function.
