+++ mbuild/__init__.py
@@ -4,6 +4,7 @@ from mbuild.coordinate_transform import *
 from mbuild.compound import *
 from mbuild.pattern import *
 from mbuild.packing import *
+from mbuild.conversion import load
 from mbuild.port import Port
 from mbuild.lattice import Lattice
 from mbuild.recipes import recipes
