+++ mbuild/compound.py
@@ -1,4 +1,4 @@
-__all__ = ['load', 'clone', 'Compound', 'Particle']
+__all__ = ['clone', 'Compound', 'Particle']
 
 from collections import OrderedDict, defaultdict, Iterable
 from copy import deepcopy
@@ -29,161 +29,8 @@ from mbuild.periodic_kdtree import PeriodicCKDTree
 from mbuild.utils.io import run_from_ipython, import_, has_networkx
 from mbuild.utils.jsutils import overwrite_nglview_default
 from mbuild.coordinate_transform import _translate, _rotate
-
-
-def load(filename_or_object, relative_to_module=None, compound=None, coords_only=False,
-         rigid=False, use_parmed=False, smiles=False, 
-         infer_hierarchy=True, **kwargs):
-    """Load a file or an existing topology into an mbuild compound.
-
-    Files are read using the MDTraj package unless the `use_parmed` argument is
-    specified as True. Please refer to http://mdtraj.org/1.8.0/load_functions.html
-    for formats supported by MDTraj and https://parmed.github.io/ParmEd/html/
-    readwrite.html for formats supported by ParmEd.
-
-    Parameters
-    ----------
-    filename_or_object : str, mdtraj.Trajectory, parmed.Structure, mbuild.Compound,
-            pybel.Molecule
-        Name of the file or topology from which to load atom and bond information.
-    relative_to_module : str, optional, default=None
-        Instead of looking in the current working directory, look for the file
-        where this module is defined. This is typically used in Compound
-        classes that will be instantiated from a different directory
-        (such as the Compounds located in mbuild.lib).
-    compound : mb.Compound, optional, default=None
-        Existing compound to load atom and bond information into.
-    coords_only : bool, optional, default=False
-        Only load the coordinates into an existing compound.
-    rigid : bool, optional, default=False
-        Treat the compound as a rigid body
-    use_parmed : bool, optional, default=False
-        Use readers from ParmEd instead of MDTraj.
-    smiles: bool, optional, default=False
-        Use Open Babel to parse filename as a SMILES string
-        or file containing a SMILES string.
-    infer_hierarchy : bool, optional, default=True
-        If True, infer hierarchy from chains and residues
-    **kwargs : keyword arguments
-        Key word arguments passed to mdTraj for loading.
-
-    Returns
-    -------
-    compound : mb.Compound
-
-    """
-    # If compound doesn't exist, we will initialize one
-    if compound is None:
-        compound = Compound()
-
-    # First check if we are loading from an existing parmed or trajectory structure
-    type_dict = {
-        pmd.Structure:compound.from_parmed,
-        md.Trajectory:compound.from_trajectory,
-    }
-    try:
-        pybel = import_('pybel')
-        type_dict.update({pybel.Molecule:compound.from_pybel})
-    except ImportError:
-        pass
-
-    if isinstance(filename_or_object, Compound):
-        return filename_or_object
-    for type in type_dict:
-        if isinstance(filename_or_object, type):
-            type_dict[type](filename_or_object,coords_only=coords_only, 
-                    infer_hierarchy=infer_hierarchy, **kwargs)
-            return compound
-    if not isinstance(filename_or_object, str):
-        raise ValueError('Input not supported.')
-
-    # Handle mbuild *.py files containing a class that wraps a structure file
-    # in its own folder. E.g., you build a system from ~/foo.py and it imports
-    # from ~/bar/baz.py where baz.py loads ~/bar/baz.pdb.
-    if relative_to_module:
-        script_path = os.path.realpath(
-            sys.modules[relative_to_module].__file__)
-        file_dir = os.path.dirname(script_path)
-        filename_or_object = os.path.join(file_dir, filename_or_object)
-
-    # Handle the case of a xyz and json file, which must use an internal reader
-    extension = os.path.splitext(filename_or_object)[-1]
-    if extension == '.json':
-        compound = compound_from_json(filename_or_object)
-        return compound
-
-    if extension == '.xyz' and not 'top' in kwargs:
-        if coords_only:
-            tmp = read_xyz(filename_or_object)
-            if tmp.n_particles != compound.n_particles:
-                raise ValueError('Number of atoms in {filename_or_object} does not match'
-                                 ' {compound}'.format(**locals()))
-            ref_and_compound = zip(tmp._particles(include_ports=False),
-                                   compound.particles(include_ports=False))
-            for ref_particle, particle in ref_and_compound:
-                particle.pos = ref_particle.pos
-        else:
-            compound = read_xyz(filename_or_object, compound=compound)
-        return compound
-
-    if extension == '.sdf':
-        pybel = import_('pybel')
-        pybel_mol = pybel.readfile('sdf', filename_or_object)
-        # pybel returns a generator, so we grab the first molecule of a list of len 1
-        # Raise ValueError user if there are more molecules
-        pybel_mol = [i for i in pybel_mol]
-        if len(pybel_mol) == 1:
-            compound.from_pybel(pybel_mol[0])
-        else:
-            compound.from_pybel(pybel_mol[0])
-            raise ValueError("More than one pybel molecule in file, more than one pybel "
-                 "molecule is not supported, using {}".format(filename_or_object))
-        return compound
-
-    if use_parmed:
-        warn(
-            "use_parmed set to True.  Bonds may be inferred from inter-particle "
-            "distances and standard residue templates!")
-        structure = pmd.load_file(filename_or_object, structure=True, **kwargs)
-        compound.from_parmed(structure, coords_only=coords_only, 
-                infer_hierarchy=infer_hierarchy)
-
-    elif smiles:
-        pybel = import_('pybel')
-        # First we try treating filename_or_object as a SMILES string
-        try:
-            mymol = pybel.readstring("smi", filename_or_object)
-        # Now we treat it as a filename
-        except(OSError, IOError):
-            # For now, we only support reading in a single smiles molecule,
-            # but pybel returns a generator, so we get the first molecule
-            # and warn the user if there is more
-
-            mymol_generator = pybel.readfile("smi", filename_or_object)
-            mymol_list = list(mymol_generator)
-            if len(mymol_list) == 1:
-                mymol = mymol_list[0]
-            else:
-                mymol = mymol_list[0]
-                warn("More than one SMILES string in file, more than one SMILES "
-                     "string is not supported, using {}".format(mymol.write("smi")))
-
-        # We create a temporary directory and mol2 file that will created from the smiles string
-        # A ParmEd structure and subsequenc mBuild compound will be created from this mol2 file
-        tmp_dir = tempfile.mkdtemp()
-        temp_file = os.path.join(tmp_dir, 'smiles_to_mol2_intermediate.mol2')
-        mymol.make3D()
-        compound = Compound()
-        compound.from_pybel(mymol, infer_hierarchy=infer_hierarchy)
-
-    else:
-        traj = md.load(filename_or_object, **kwargs)
-        compound.from_trajectory(traj, frame=-1, coords_only=coords_only,
-                infer_hierarchy=infer_hierarchy)
-
-    if rigid:
-        compound.label_rigid_bodies()
-    return compound
+from mbuild import conversion
+from mbuild.conversion import load
 
 
 def clone(existing_compound, clone_of=None, root_container=None):
@@ -2055,50 +1902,13 @@ class Compound(object):
             Only read coordinate information
         infer_hierarchy : bool, optional, default=True
             If True, infer compound hierarchy from chains and residues
+
+        See also
+        --------
+        mbuild.conversion.from_trajectory
         """
-        if coords_only:
-            if traj.n_atoms != self.n_particles:
-                raise ValueError('Number of atoms in {traj} does not match'
-                                 ' {self}'.format(**locals()))
-            atoms_particles = zip(traj.topology.atoms,
-                                  self.particles(include_ports=False))
-            if None in self._particles(include_ports=False):
-                raise ValueError('Some particles are None')
-            for mdtraj_atom, particle in atoms_particles:
-                particle.pos = traj.xyz[frame, mdtraj_atom.index]
-            return
-
-        atom_mapping = dict()
-        for chain in traj.topology.chains:
-            if traj.topology.n_chains > 1:
-                chain_compound = Compound()
-                self.add(chain_compound, 'chain[$]')
-            else:
-                chain_compound = self
-            for res in chain.residues:
-                if infer_hierarchy:
-                    res_compound = Compound(name=res.name)
-                    chain_compound.add(res_compound)
-                    parent_cmpd = res_compound
-                else:
-                    parent_cmpd = chain_compound
-                for atom in res.atoms:
-                    new_atom = Particle(name=str(atom.name),
-                                        pos=traj.xyz[frame, atom.index])
-                    parent_cmpd.add(
-                        new_atom, label='{0}[$]'.format(
-                            atom.name))
-                    atom_mapping[atom] = new_atom
-
-        for mdtraj_atom1, mdtraj_atom2 in traj.topology.bonds:
-            atom1 = atom_mapping[mdtraj_atom1]
-            atom2 = atom_mapping[mdtraj_atom2]
-            self.add_bond((atom1, atom2))
-
-        if np.any(traj.unitcell_lengths) and np.any(traj.unitcell_lengths[0]):
-            self.periodicity = traj.unitcell_lengths[0]
-        else:
-            self.periodicity = np.array([0., 0., 0.])
+        conversion.from_trajectory(traj=traj, compound=self, frame=frame,
+            coords_only=coords_only, infer_hierarchy=True)
 
     def to_trajectory(self, show_ports=False, chains=None,
                       residues=None, box=None):
@@ -2126,153 +1936,11 @@ class Compound(object):
 
         See also
         --------
-        _to_topology
+        mbuild.conversion.to_trajectory
 
         """
-        atom_list = [particle for particle in self.particles(show_ports)]
-
-        top = self._to_topology(atom_list, chains, residues)
-
-        # Coordinates.
-        xyz = np.ndarray(shape=(1, top.n_atoms, 3), dtype='float')
-        for idx, atom in enumerate(atom_list):
-            xyz[0, idx] = atom.pos
-
-        # Unitcell information.
-        unitcell_angles = [90.0, 90.0, 90.0]
-        if box is None:
-            unitcell_lengths = np.empty(3)
-            for dim, val in enumerate(self.periodicity):
-                if val:
-                    unitcell_lengths[dim] = val
-                else:
-                    unitcell_lengths[dim] = self.boundingbox.lengths[dim] + 0.5
-        else:
-            unitcell_lengths = box.lengths
-            unitcell_angles = box.angles
-
-        return md.Trajectory(xyz, top, unitcell_lengths=unitcell_lengths,
-                             unitcell_angles=unitcell_angles)
-
-    def _to_topology(self, atom_list, chains=None, residues=None):
-        """Create a mdtraj.Topology from a Compound.
-
-        Parameters
-        ----------
-        atom_list : list of mb.Compound
-            Atoms to include in the topology
-        chains : mb.Compound or list of mb.Compound
-            Chain types to add to the topology
-        residues : str of list of str
-            Labels of residues in the Compound. Residues are assigned by
-            checking against Compound.name.
-
-        Returns
-        -------
-        top : mdtraj.Topology
-
-        See Also
-        --------
-        mdtraj.Topology : Details on the mdtraj Topology object
-
-        """
-        from mdtraj.core.topology import Topology
-
-        if isinstance(chains, str):
-            chains = [chains]
-        if isinstance(chains, (list, set)):
-            chains = tuple(chains)
-
-        if isinstance(residues, str):
-            residues = [residues]
-        if isinstance(residues, (list, set)):
-            residues = tuple(residues)
-        top = Topology()
-        atom_mapping = {}
-
-        default_chain = top.add_chain()
-        default_residue = top.add_residue('RES', default_chain)
-
-        compound_residue_map = dict()
-        atom_residue_map = dict()
-        compound_chain_map = dict()
-        atom_chain_map = dict()
-
-        for atom in atom_list:
-            # Chains
-            if chains:
-                if atom.name in chains:
-                    current_chain = top.add_chain()
-                    compound_chain_map[atom] = current_chain
-                else:
-                    for parent in atom.ancestors():
-                        if chains and parent.name in chains:
-                            if parent not in compound_chain_map:
-                                current_chain = top.add_chain()
-                                compound_chain_map[parent] = current_chain
-                                current_residue = top.add_residue(
-                                    'RES', current_chain)
-                            break
-                    else:
-                        current_chain = default_chain
-            else:
-                current_chain = default_chain
-            atom_chain_map[atom] = current_chain
-
-            # Residues
-            if residues:
-                if atom.name in residues:
-                    current_residue = top.add_residue(atom.name, current_chain)
-                    compound_residue_map[atom] = current_residue
-                else:
-                    for parent in atom.ancestors():
-                        if residues and parent.name in residues:
-                            if parent not in compound_residue_map:
-                                current_residue = top.add_residue(
-                                    parent.name, current_chain)
-                                compound_residue_map[parent] = current_residue
-                            break
-                    else:
-                        current_residue = default_residue
-            else:
-                if chains:
-                    try:  # Grab the default residue from the custom chain.
-                        current_residue = next(current_chain.residues)
-                    except StopIteration:  # Add the residue to the current chain
-                        current_residue = top.add_residue('RES', current_chain)
-                else:  # Grab the default chain's default residue
-                    current_residue = default_residue
-            atom_residue_map[atom] = current_residue
-
-            # Add the actual atoms
-            try:
-                elem = get_by_symbol(atom.name)
-            except KeyError:
-                elem = get_by_symbol("VS")
-            at = top.add_atom(atom.name, elem, atom_residue_map[atom])
-            at.charge = atom.charge
-            atom_mapping[atom] = at
-
-        # Remove empty default residues.
-        chains_to_remove = [
-            chain for chain in top.chains if chain.n_atoms == 0]
-        residues_to_remove = [res for res in top.residues if res.n_atoms == 0]
-        for chain in chains_to_remove:
-            top._chains.remove(chain)
-        for res in residues_to_remove:
-            for chain in top.chains:
-                try:
-                    chain._residues.remove(res)
-                except ValueError:  # Already gone.
-                    pass
-
-        for atom1, atom2 in self.bonds():
-            # Ensure that both atoms are part of the compound. This becomes an
-            # issue if you try to convert a sub-compound to a topology which is
-            # bonded to a different subcompound.
-            if all(a in atom_mapping.keys() for a in [atom1, atom2]):
-                top.add_bond(atom_mapping[atom1], atom_mapping[atom2])
-        return top
+        return conversion.to_trajectory(compound=self, show_ports=show_ports,
+            chains=chains, residues=residues, box=box)
 
     def from_parmed(self, structure, coords_only=False,
             infer_hierarchy=True):
@@ -2290,59 +1958,8 @@ class Compound(object):
         infer_hierarchy : bool, optional, default=True
             If true, infer compound hierarchy from chains and residues
         """
-        if coords_only:
-            if len(structure.atoms) != self.n_particles:
-                raise ValueError(
-                    'Number of atoms in {structure} does not match'
-                    ' {self}'.format(
-                        **locals()))
-            atoms_particles = zip(structure.atoms,
-                                  self.particles(include_ports=False))
-            if None in self._particles(include_ports=False):
-                raise ValueError('Some particles are None')
-            for parmed_atom, particle in atoms_particles:
-                particle.pos = np.array([parmed_atom.xx,
-                                         parmed_atom.xy,
-                                         parmed_atom.xz]) / 10
-            return
-
-        atom_mapping = dict()
-        chain_id = None
-        chains = defaultdict(list)
-        for residue in structure.residues:
-            chains[residue.chain].append(residue)
-
-        for chain, residues in chains.items():
-            if len(chains) > 1:
-                chain_compound = Compound()
-                self.add(chain_compound, chain_id)
-            else:
-                chain_compound = self
-            for residue in residues:
-                if infer_hierarchy:
-                    residue_compound = Compound(name=residue.name)
-                    chain_compound.add(residue_compound)
-                    parent_cmpd = residue_compound
-                else:
-                    parent_cmpd = chain_compound
-                for atom in residue.atoms:
-                    pos = np.array([atom.xx, atom.xy, atom.xz]) / 10
-                    new_atom = Particle(name=str(atom.name), pos=pos)
-                    parent_cmpd.add(
-                        new_atom, label='{0}[$]'.format(
-                            atom.name))
-                    atom_mapping[atom] = new_atom
-
-        for bond in structure.bonds:
-            atom1 = atom_mapping[bond.atom1]
-            atom2 = atom_mapping[bond.atom2]
-            self.add_bond((atom1, atom2))
-
-        if structure.box is not None:
-            # Convert from A to nm
-            self.periodicity = 0.1 * structure.box[0:3]
-        else:
-            self.periodicity = np.array([0., 0., 0.])
+        conversion.from_parmed(structure=structure,compound=self,
+            coords_only=coords_only, infer_hierarchy=infer_hierarchy)
 
     def to_parmed(self, box=None, title='', residues=None, show_ports=False,
             infer_residues=False):
@@ -2373,120 +1990,12 @@ class Compound(object):
 
         See Also
         --------
+        mbuild.conversion.to_parmed
         parmed.structure.Structure : Details on the ParmEd Structure object
 
         """
-        structure = pmd.Structure()
-        structure.title = title if title else self.name
-        atom_mapping = {}  # For creating bonds below
-        guessed_elements = set()
-
-        # Attempt to grab residue names based on names of children
-        if not residues and infer_residues:
-            residues = list(set([child.name for child in self.children]))
-
-        if isinstance(residues, str):
-            residues = [residues]
-        if isinstance(residues, (list, set)):
-            residues = tuple(residues)
-
-        default_residue = pmd.Residue('RES')
-        port_residue = pmd.Residue('PRT')
-        compound_residue_map = dict()
-        atom_residue_map = dict()
-
-        # Loop through particles and add initialize ParmEd atoms
-        for atom in self.particles(include_ports=show_ports):
-            if atom.port_particle:
-                current_residue = port_residue
-                atom_residue_map[atom] = current_residue
-
-                if current_residue not in structure.residues:
-                    structure.residues.append(current_residue)
-
-                pmd_atom = pmd.Atom(atomic_number=0, name='VS',
-                                    mass=0, charge=0)
-                pmd_atom.xx, pmd_atom.xy, pmd_atom.xz = atom.pos * 10  # Angstroms
-
-            else:
-                if residues and atom.name in residues:
-                    current_residue = pmd.Residue(atom.name)
-                    atom_residue_map[atom] = current_residue
-                    compound_residue_map[atom] = current_residue
-                elif residues:
-                    for parent in atom.ancestors():
-                        if residues and parent.name in residues:
-                            if parent not in compound_residue_map:
-                                current_residue = pmd.Residue(parent.name)
-                                compound_residue_map[parent] = current_residue
-                            atom_residue_map[atom] = current_residue
-                            break
-                    else:  # Did not find specified residues in ancestors.
-                        current_residue = default_residue
-                        atom_residue_map[atom] = current_residue
-                else:
-                    current_residue = default_residue
-                    atom_residue_map[atom] = current_residue
-
-                if current_residue not in structure.residues:
-                    structure.residues.append(current_residue)
-
-                atomic_number = None
-                name = ''.join(char for char in atom.name if not char.isdigit())
-                try:
-                    atomic_number = AtomicNum[atom.name.capitalize()]
-                except KeyError:
-                    element = element_by_name(atom.name.capitalize())
-                    if name not in guessed_elements:
-                        warn(
-                            'Guessing that "{}" is element: "{}"'.format(
-                                atom, element))
-                        guessed_elements.add(name)
-                else:
-                    element = atom.name.capitalize()
-
-                atomic_number = atomic_number or AtomicNum[element]
-                mass = Mass[element]
-                pmd_atom = pmd.Atom(atomic_number=atomic_number, name=atom.name,
-                                    mass=mass, charge=atom.charge)
-                pmd_atom.xx, pmd_atom.xy, pmd_atom.xz = atom.pos * 10  # Angstroms
-
-            residue = atom_residue_map[atom]
-            structure.add_atom(pmd_atom, resname=residue.name,
-                               resnum=residue.idx)
-
-            atom_mapping[atom] = pmd_atom
-
-        # "Claim" all of the items it contains and subsequently index all of its items
-        structure.residues.claim()
-
-        # Create and add bonds to ParmEd Structure
-        for atom1, atom2 in self.bonds():
-            bond = pmd.Bond(atom_mapping[atom1], atom_mapping[atom2])
-            structure.bonds.append(bond)
-        # pad box with .25nm buffers
-        if box is None:
-            box = self.boundingbox
-            box_vec_max = box.maxs.tolist()
-            box_vec_min = box.mins.tolist()
-            for dim, val in enumerate(self.periodicity):
-                if val:
-                    box_vec_max[dim] = val
-                    box_vec_min[dim] = 0.0
-                if not val:
-                    box_vec_max[dim] += 0.25
-                    box_vec_min[dim] -= 0.25
-            box = Box(mins=box_vec_min, maxs=box_vec_max)
-
-        box_vector = np.empty(6)
-        if box.angles is not None:
-            box_vector[3:6] = box.angles
-        else:
-            box_vector[3] = box_vector[4] = box_vector[5] = 90.0
-        for dim in range(3):
-            box_vector[dim] = box.lengths[dim] * 10
-        structure.box = box_vector
-        return structure
+        return conversion.to_parmed(compound=self, box=box, title=title,
+            residues=residues, show_ports=show_ports, infer_residues=infer_residues)
 
     def to_networkx(self, names_only=False):
         """Create a NetworkX graph representing the hierarchy of a Compound.
@@ -2509,38 +2018,10 @@ class Compound(object):
 
         See Also
         --------
+        mbuild.conversion.to_networkx
         mbuild.bond_graph
         """
-        nx = import_('networkx')
-
-        nodes = list()
-        edges = list()
-        if names_only:
-            nodes.append(self.name + '_' + str(id(self)))
-        else:
-            nodes.append(self)
-        nodes, edges = self._iterate_children(nodes, edges, names_only=names_only)
-
-        graph = nx.DiGraph()
-        graph.add_nodes_from(nodes)
-        graph.add_edges_from(edges)
-        return graph
-
-    def _iterate_children(self, nodes, edges, names_only=False):
-        """ Create nodes and edges that connect parents and their corresponding children"""
-        if not self.children:
-            return nodes, edges
-        for child in self.children:
-            if names_only:
-                unique_name = child.name + '_' + str(id(child))
-                unique_name_parent = child.parent.name + '_' + str((id(child.parent)))
-                nodes.append(unique_name)
-                edges.append([unique_name_parent, unique_name])
-            else:
-                nodes.append(child)
-                edges.append([child.parent, child])
-            nodes, edges = child._iterate_children(nodes, edges, names_only=names_only)
-        return nodes, edges
+        return conversion.to_networkx(compound=self, names_only=names_only)
 
     def to_pybel(self, box=None, title='', residues=None, show_ports=False,
             infer_residues=False):
@@ -2563,114 +2044,20 @@ class Compound(object):
         ------
         pybel.Molecule
 
+        See also
+        --------
+        mbuild.conversion.to_pybel
+
         Notes
         -----
         Most of the mb.Compound is first converted to openbabel.OBMol
         And then pybel creates a pybel.Molecule from the OBMol
         Bond orders are assumed to be 1
         OBMol atom indexing starts at 1, with spatial dimension Angstrom
+
         """
-
-        openbabel = import_('openbabel')
-        pybel = import_('pybel')
-
-        mol = openbabel.OBMol()
-        particle_to_atom_index = {}
-
-        if not residues and infer_residues:
-            residues = list(set([child.name for child in self.children]))
-        if isinstance(residues, str):
-            residues = [residues]
-        if isinstance(residues, (list, set)):
-            residues = tuple(residues)
-
-        compound_residue_map = dict()
-        atom_residue_map = dict()
-
-        for i, part in enumerate(self.particles(include_ports=show_ports)):
-            if residues and part.name in residues:
-                current_residue = mol.NewResidue()
-                current_residue.SetName(part.name)
-                atom_residue_map[part] = current_residue
-                compound_residue_map[part] = current_residue
-            elif residues:
-                for parent in part.ancestors():
-                    if residues and parent.name in residues:
-                        if parent not in compound_residue_map:
-                            current_residue = mol.NewResidue()
-                            current_residue.SetName(parent.name)
-                            compound_residue_map[parent] = current_residue
-                        atom_residue_map[part] = current_residue
-                        break
-                else:  # Did not find specified residues in ancestors.
-                    current_residue = mol.NewResidue()
-                    current_residue.SetName("RES")
-                    atom_residue_map[part] = current_residue
-            else:
-                current_residue = mol.NewResidue()
-                current_residue.SetName("RES")
-                atom_residue_map[part] = current_residue
-
-            temp = mol.NewAtom()
-            residue = atom_residue_map[part]
-            temp.SetResidue(residue)
-            if part.port_particle:
-                temp.SetAtomicNum(0)
-            else:
-                try:
-                    temp.SetAtomicNum(AtomicNum[part.name.capitalize()])
-                except KeyError:
-                    warn("Could not infer atomic number from "
-                            "{}, setting to 0".format(part.name))
-                    temp.SetAtomicNum(0)
-
-
-            temp.SetVector(*(part.xyz[0]*10))
-            particle_to_atom_index[part] = i
-
-        ucell = openbabel.OBUnitCell()
-        if box is None:
-            box = self.boundingbox
-        a, b, c = 10.0 * box.lengths
-        alpha, beta, gamma = np.radians(box.angles)
-
-        cosa = np.cos(alpha)
-        cosb = np.cos(beta)
-        sinb = np.sin(beta)
-        cosg = np.cos(gamma)
-        sing = np.sin(gamma)
-        mat_coef_y = (cosa - cosb * cosg) / sing
-        mat_coef_z = np.power(sinb, 2, dtype=float) - \
-                    np.power(mat_coef_y, 2, dtype=float)
-
-        if mat_coef_z > 0.:
-            mat_coef_z = np.sqrt(mat_coef_z)
-        else:
-            raise Warning('Non-positive z-vector. Angles {} '
-                                  'do not generate a box with the z-vector in the'
-                                  'positive z direction'.format(box.angles))
-
-        box_vec = [[1, 0, 0],
-                    [cosg, sing, 0],
-                    [cosb, mat_coef_y, mat_coef_z]]
-        box_vec = np.asarray(box_vec)
-        box_mat = (np.array([a,b,c])* box_vec.T).T
-        first_vector = openbabel.vector3(*box_mat[0])
-        second_vector = openbabel.vector3(*box_mat[1])
-        third_vector = openbabel.vector3(*box_mat[2])
-        ucell.SetData(first_vector, second_vector, third_vector)
-        mol.CloneData(ucell)
-
-        for bond in self.bonds():
-            bond_order = 1
-            mol.AddBond(particle_to_atom_index[bond[0]]+1,
-                    particle_to_atom_index[bond[1]]+1,
-                    bond_order)
-
-        pybelmol = pybel.Molecule(mol)
-        pybelmol.title = title if title else self.name
-
-        return pybelmol
+        return conversion.to_pybel(compound=self, box=box, title=title,
+            residues=residues, show_ports=show_ports)
 
     def from_pybel(self, pybel_mol, use_element=True, coords_only=False,
             infer_hierarchy=True):
@@ -2689,64 +2076,13 @@ class Compound(object):
         infer_hierarchy : bool, optional, default=True
             If True, infer hierarchy from residues
 
+        See also
+        --------
+        mbuild.conversion.from_pybel
+
         """
-        openbabel = import_("openbabel")
-        self.name = pybel_mol.title.split('.')[0]
-        resindex_to_cmpd = {}
-
-        if coords_only:
-            raise Warning('coords_only=True not yet implemented for '
-                    'conversion from pybel')
-        # Iterating through pybel_mol for atom/residue information
-        # This could just as easily be implemented by
-        # an OBMolAtomIter from the openbabel library,
-        # but this seemed more convenient at time of writing
-        # pybel atoms are 1-indexed, coordinates in Angstrom
-        for atom in pybel_mol.atoms:
-            xyz = np.array(atom.coords)/10
-            if use_element:
-                try:
-                    temp_name = Element[atom.atomicnum]
-                except KeyError:
-                    warn("No element detected for atom at index "
-                            "{} with number {}, type {}".format(
-                                atom.idx, atom.atomicnum, atom.type))
-                    temp_name = atom.type
-            else:
-                temp_name = atom.type
-            temp = Particle(name=temp_name, pos=xyz)
-            if infer_hierarchy and hasattr(atom, 'residue'): 
-                # Is there a safer way to check for res?
-                if atom.residue.idx not in resindex_to_cmpd:
-                    res_cmpd = Compound(name=atom.residue.name)
-                    resindex_to_cmpd[atom.residue.idx] = res_cmpd
-                    self.add(res_cmpd)
-                resindex_to_cmpd[atom.residue.idx].add(temp)
-            else:
-                self.add(temp)
-
-        # Iterating through pybel_mol.OBMol for bond information
-        # Bonds are 0-indexed, but the atoms are 1-indexed
-        # Bond information doesn't appear stored in pybel_mol,
-        # so we need to look into the OBMol object,
-        # using an iterator from the openbabel library
-        for bond in openbabel.OBMolBondIter(pybel_mol.OBMol):
-            self.add_bond([self[bond.GetBeginAtomIdx()-1],
-                            self[bond.GetEndAtomIdx()-1]])
-
-        if hasattr(pybel_mol, 'unitcell'):
-            box = Box(lengths=[pybel_mol.unitcell.GetA()/10,
-                                pybel_mol.unitcell.GetB()/10,
-                                pybel_mol.unitcell.GetC()/10],
-                        angles=[pybel_mol.unitcell.GetAlpha(),
-                                pybel_mol.unitcell.GetBeta(),
-                                pybel_mol.unitcell.GetGamma()])
-            self.periodicity = box.lengths
-        else:
-            warn("No unitcell detected for pybel.Molecule {}".format(pybel_mol))
-#       TODO: Decide how to gather PBC information from openbabel. Options may
-#             include storing it in .periodicity or writing a separate function
-#             that returns the box.
+        conversion.from_pybel(pybel_mol=pybel_mol, compound=self,
+            use_element=use_element, coords_only=coords_only)
 
     def to_intermol(self, molecule_types=None): # pragma: no cover
         """Create an InterMol system from a Compound.
@@ -2758,46 +2094,13 @@ class Compound(object):
         Returns
         -------
         intermol_system : intermol.system.System
+
+        See also
+        --------
+        mbuild.conversion.to_intermol
+
         """
-        from intermol.atom import Atom as InterMolAtom
-        from intermol.molecule import Molecule
-        from intermol.system import System
-        import simtk.unit as u
-
-        if isinstance(molecule_types, list):
-            molecule_types = tuple(molecule_types)
-        elif molecule_types is None:
-            molecule_types = (type(self),)
-        intermol_system = System()
-
-        last_molecule_compound = None
-        for atom_index, atom in enumerate(self.particles()):
-            for parent in atom.ancestors():
-                # Don't want inheritance via isinstance().
-                if type(parent) in molecule_types:
-                    # Check if we have encountered this molecule type before.
-                    if parent.name not in intermol_system.molecule_types:
-                        self._add_intermol_molecule_type(
-                            intermol_system, parent)
-                    if parent != last_molecule_compound:
-                        last_molecule_compound = parent
-                        last_molecule = Molecule(name=parent.name)
-                        intermol_system.add_molecule(last_molecule)
-                    break
-            else:
-                # Should never happen if molecule_types only contains
-                # type(self)
-                raise ValueError(
-                    'Found an atom {} that is not part of any of '
-                    'the specified molecule types {}'.format(
-                        atom, molecule_types))
-
-            # Add the actual intermol atoms.
-            intermol_atom = InterMolAtom(atom_index + 1, name=atom.name,
-                                         residue_index=1, residue_name='RES')
-            intermol_atom.position = atom.pos * u.nanometers
-            last_molecule.add_atom(intermol_atom)
-        return intermol_system
+        return conversion.to_intermol(compound=self, molecule_types=None)
 
     def get_smiles(self):
         """Get SMILES string for compound
@@ -2815,27 +2118,6 @@ class Compound(object):
         # we only need the smiles string
         smiles = pybel_cmp.write().split()[0]
         return smiles
-
-
-    @staticmethod
-    def _add_intermol_molecule_type(intermol_system, parent):  # pragma: no cover
-        """Create a molecule type for the parent and add bonds.
-
-        This method takes an intermol system and adds a
-        parent compound, including its particles and bonds, to it.
-        """
-        from intermol.moleculetype import MoleculeType
-        from intermol.forces.bond import Bond as InterMolBond
-
-        molecule_type = MoleculeType(name=parent.name)
-        intermol_system.add_molecule_type(molecule_type)
-
-        for index, parent_atom in enumerate(parent.particles()):
-            parent_atom.index = index + 1
-
-        for atom1, atom2 in parent.bonds():
-            intermol_bond = InterMolBond(atom1.index, atom2.index)
-            molecule_type.bonds.add(intermol_bond)
 
     def __getitem__(self, selection):
         if isinstance(selection, int):
