+++ mbuild/tests/test_compound.py
@@ -872,7 +872,7 @@ class TestCompound(BaseTest):
     def test_load_mol2_mdtraj(self):
         with pytest.raises(KeyError):
             mb.load(get_fn('benzene-nonelement.mol2'))
-        mb.load(get_fn('benzene-nonelement.mol2'), use_parmed=True)
+        mb.load(get_fn('benzene-nonelement.mol2'), backend='parmed')
 
     def test_siliane_bond_number(self, silane):
         assert silane.n_bonds == 4
